requirements.txt
numpy==1.15.4
opencv-python==3.4.4.19
Pillow==5.3.0
PyAutoGUI==0.9.38
PyMsgBox==1.0.6
pynput==1.4
pyperclip==1.7.0
PyScreeze==0.1.18
pytesseract==0.2.5
PyTweening==1.0.3
six==1.12.0
